+++ packages/material_leap/lib/src/helpers/breakpoints.dart
@@ -1,8 +1,8 @@
 class LeapBreakpoints {
-  static const int compact = 600;
-  static const int medium = 840;
-  static const int expanded = 1200;
-  static const int large = 1600;
+  static const double compact = 600;
+  static const double medium = 840;
+  static const double expanded = 1200;
+  static const double large = 1600;
 
   const LeapBreakpoints._();
 }
